+++ backend/api/views/compute_plan_graph.py
@@ -1,13 +1,16 @@
 import structlog
-from django.contrib.postgres.expressions import ArraySubquery
+from django.contrib.postgres.aggregates import JSONBAgg
 from django.db.models import Field
 from django.db.models import Func
 from django.db.models import OuterRef
+from django.db.models import Subquery
 from django.db.models.expressions import F
+from django.db.models.expressions import Value
 from rest_framework import status
 from rest_framework.decorators import api_view
 
 from api.models import ComputeTask
+from api.models import ComputeTaskInput
 from api.views.sql_utils import Any
 from api.views.utils import ApiResponse
 from api.views.utils import get_channel_name
@@ -18,121 +21,63 @@ Field.register_lookup(Any)
 
 logger = structlog.get_logger(__name__)
 
-TASK_CATEGORY_INPUTS = {
-    ComputeTask.Category.TASK_TRAIN: [{"id": "in/model", "kind": "model"}],
-    ComputeTask.Category.TASK_TEST: [{"id": "in/predictions", "kind": "model"}],
-    ComputeTask.Category.TASK_COMPOSITE: [
-        {"id": "in/head_model", "kind": "model"},
-        {"id": "in/trunk_model", "kind": "model"},
-    ],
-    ComputeTask.Category.TASK_AGGREGATE: [{"id": "in/models[]", "kind": "model"}],
-    ComputeTask.Category.TASK_PREDICT: [{"id": "in/tested_model", "kind": "model"}],
-}
-
-TASK_CATEGORY_OUTPUTS = {
-    ComputeTask.Category.TASK_TRAIN: [{"id": "out/model", "kind": "model"}],
-    ComputeTask.Category.TASK_TEST: [{"id": "out/perf", "kind": "performance"}],
-    ComputeTask.Category.TASK_COMPOSITE: [
-        {"id": "out/head_model", "kind": "model"},
-        {"id": "out/trunk_model", "kind": "model"},
-    ],
-    ComputeTask.Category.TASK_AGGREGATE: [{"id": "out/model", "kind": "model"}],
-    ComputeTask.Category.TASK_PREDICT: [{"id": "out/predictions", "kind": "model"}],
-}
-
 MAX_TASKS_DISPLAYED = 1000
 
 
-def _get_task_inputs(category):
-    return TASK_CATEGORY_INPUTS.get(category, [])
-
-
-def _get_task_outputs(category):
-    return TASK_CATEGORY_OUTPUTS.get(category, [])
-
-
-def _get_target_input(edge):
-    target_category = edge.get("target_task_category")
-    inputs = [input["id"] for input in TASK_CATEGORY_INPUTS.get(target_category, [])]
-    if TASK_CATEGORY_INPUTS.get(target_category) is None:
-        raise Exception(
-            (
-                "Failed to build CP workflow graph. Could not build an edge for: "
-                f"task {edge['target_task_key']} has a category {target_category}."
-            )
-        )
-    if (
-        target_category == ComputeTask.Category.TASK_COMPOSITE
-        and edge.get("source_task_category") == ComputeTask.Category.TASK_AGGREGATE
-    ):
-        input = inputs[1]
-    else:
-        input = inputs[0]
-
-    return input
-
-
-def _get_source_output(edge):
-    source_category = edge.get("source_task_category")
-    outputs = [output["id"] for output in TASK_CATEGORY_OUTPUTS.get(source_category, [])]
-    if TASK_CATEGORY_OUTPUTS.get(source_category) is None:
-        raise Exception(
-            (
-                "Failed to build CP workflow graph. Could not build an edge for: "
-                f"task {edge['target_task_key']} has a category {source_category}."
-            )
-        )
-    if (
-        source_category == ComputeTask.Category.TASK_COMPOSITE
-        and edge.get("target_task_category") == ComputeTask.Category.TASK_AGGREGATE
-    ):
-        output = outputs[1]
-    else:
-        output = outputs[0]
-
-    return output
+class JsonbBuildObj(Func):
+    function = "jsonb_build_object"
 
 
 @api_view(["GET"])
 def get_cp_graph(request, compute_plan_pk):
-    """Return a workflow graph for each task of the computeplan"""
+    """Return a workflow graph for each task of the computeplan.
+    As this view is used to build a graph and therefore cannot be paginated yet,
+    it does not use any serialization for performance optimization"""
     validate_key(compute_plan_pk)
 
-    tasks = ComputeTask.objects.filter(compute_plan__key=compute_plan_pk, channel=get_channel_name(request)).values(
-        "key",
-        "rank",
-        "worker",
-        "status",
-        "category",
+    outputs = (
+        ComputeTask.objects.filter(key=OuterRef("pk"))
+        .annotate(
+            outputs_specs=JSONBAgg(
+                JsonbBuildObj(Value("identifier"), F("algo__outputs__identifier"), Value("kind"), "algo__outputs__kind")
+            ),
+        )
+        .values("outputs_specs")
+    )
+
+    inputs = (
+        ComputeTask.objects.filter(key=OuterRef("pk"))
+        .annotate(
+            inputs_specs=JSONBAgg(
+                JsonbBuildObj(Value("identifier"), F("algo__inputs__identifier"), Value("kind"), "algo__inputs__kind")
+            ),
+        )
+        .values("inputs_specs")
+    )
+
+    tasks = (
+        ComputeTask.objects.filter(compute_plan__key=compute_plan_pk, channel=get_channel_name(request))
+        .annotate(inputs_specs=Subquery(inputs), outputs_specs=Subquery(outputs))
+        .values("key", "rank", "worker", "status", "inputs_specs", "outputs_specs")
     )
 
     # Set a task limitation for performances issues
-    if tasks.count() > MAX_TASKS_DISPLAYED:
+    if len(tasks) > MAX_TASKS_DISPLAYED:
         return ApiResponse(
             data={"message": f"Cannot build workflow graph for more than {MAX_TASKS_DISPLAYED} tasks."},
             status=status.HTTP_403_FORBIDDEN,
         )
 
-    for task in tasks:
-        task["inputs"] = _get_task_inputs(task.get("category"))
-        task["outputs"] = _get_task_outputs(task.get("category"))
+    edges_qs = ComputeTaskInput.objects.filter(
+        task__compute_plan__key=compute_plan_pk, channel=get_channel_name(request)
+    ).exclude(parent_task_key__isnull=True, parent_task_output_identifier__isnull=True)
 
-    parent_tasks = ComputeTask.objects.filter(key__any=OuterRef("parent_tasks"))
-
-    edges = (
-        ComputeTask.objects.filter(compute_plan__key=compute_plan_pk, channel=get_channel_name(request))
-        .annotate(
-            source_task_key=Func(ArraySubquery(parent_tasks.values("key")), function="unnest"),
-            source_task_category=Func(ArraySubquery(parent_tasks.values("category")), function="unnest"),
-            target_task_key=F("key"),
-            target_task_category=F("category"),
-        )
-        .values("source_task_key", "target_task_key", "source_task_category", "target_task_category")
-    )
-
-    for edge in edges:
-        edge["source_output_name"] = _get_source_output(edge)
-        edge["target_input_name"] = _get_target_input(edge)
+    edges = edges_qs.annotate(
+        source_task_key=F("parent_task_key"),
+        source_output_identifier=F("parent_task_output_identifier"),
+        target_task_key=F("task__key"),
+        target_input_identifier=F("identifier"),
+    ).values("source_task_key", "source_output_identifier", "target_task_key", "target_input_identifier")
 
     return ApiResponse(
         data={